requirements.txt
# === Interface principale ===
streamlit>=1.36.0

# === Traitement et rendu graphique ===
numpy>=1.26.0
pillow>=10.3.0

# === Visualisation et affichage ===
matplotlib>=3.8.0
colorama>=0.4.6
pygments>=2.17.0

# === Parsing et manipulation binaire ===
bitstring>=4.2.3
construct>=2.10.70
